Team4_code/requirement.txt
Environment:
Python 3

Packages needed:
flask 
yfinance
pandas 
matplotlib
xlsxwriter
TA-lib
openai
google-cloud-vision
dateutil
tkinter
pillow
request



